use Scroll canonical bridge for WETH transfers
