+++ package.json
@@ -10,7 +10,7 @@
     "node": ">=20"
   },
   "dependencies": {
-    "@across-protocol/constants": "^3.1.10",
+    "@across-protocol/constants": "^3.1.11",
     "@across-protocol/contracts": "^3.0.6",
     "@across-protocol/sdk": "^3.1.13",
     "@arbitrum/sdk": "^3.1.3",
