+++ src/adapter/bridges/ScrollERC20Bridge.ts
@@ -1,10 +1,22 @@
-import { Contract, BigNumber, paginatedEventQuery, EventSearchConfig, Signer, Provider } from "../../utils";
+import {
+  Contract,
+  BigNumber,
+  paginatedEventQuery,
+  EventSearchConfig,
+  Signer,
+  Provider,
+  toWei,
+  fixedPointAdjustment,
+} from "../../utils";
 import { CONTRACT_ADDRESSES } from "../../common";
 import { BaseBridgeAdapter, BridgeTransactionDetails, BridgeEvents } from "./BaseBridgeAdapter";
 import { processEvent } from "../utils";
 
 export class ScrollERC20Bridge extends BaseBridgeAdapter {
-  protected l2Gas = 20000;
+  // Gas limit obtained here: https://docs.scroll.io/en/developers/l1-and-l2-bridging/eth-and-erc20-token-bridge
+  protected l2Gas = 200000;
+  protected feeMultiplier = toWei(1.5);
+  protected readonly scrollGasPriceOracle: Contract;
   constructor(
     l2chainId: number,
     hubChainId: number,
@@ -16,10 +28,13 @@ export class ScrollERC20Bridge extends BaseBridgeAdapter {
     _l1Token;
     const { address: l1Address, abi: l1Abi } = CONTRACT_ADDRESSES[hubChainId].scrollGatewayRouter;
     const { address: l2Address, abi: l2Abi } = CONTRACT_ADDRESSES[l2chainId].scrollGatewayRouter;
+    const { address: gasPriceOracleAddress, abi: gasPriceOracleAbi } =
+      CONTRACT_ADDRESSES[hubChainId].scrollGasPriceOracle;
     super(l2chainId, hubChainId, l1Signer, l2SignerOrProvider, [l1Address]);
 
     this.l1Bridge = new Contract(l1Address, l1Abi, l1Signer);
     this.l2Bridge = new Contract(l2Address, l2Abi, l2SignerOrProvider);
+    this.scrollGasPriceOracle = new Contract(gasPriceOracleAddress, gasPriceOracleAbi, l1Signer);
   }
 
   async constructL1ToL2Txn(
@@ -28,10 +43,13 @@ export class ScrollERC20Bridge extends BaseBridgeAdapter {
     l2Token: string,
     amount: BigNumber
   ): Promise<BridgeTransactionDetails> {
+    const baseFee = await this.getScrollGasPriceOracle().estimateCrossDomainMessageFee(this.l2Gas);
+    const bufferedFee = baseFee.mul(this.feeMultiplier).div(fixedPointAdjustment);
     return Promise.resolve({
       contract: this.getL1Bridge(),
       method: "depositERC20",
       args: [l1Token, toAddress, amount, this.l2Gas],
+      value: bufferedFee,
     });
   }
 
@@ -67,5 +85,9 @@ export class ScrollERC20Bridge extends BaseBridgeAdapter {
     return {
       [this.resolveL2TokenAddress(l1Token)]: events.map((event) => processEvent(event, "amount", "to", "from")),
     };
+  }
+
+  protected getScrollGasPriceOracle(): Contract {
+    return this.scrollGasPriceOracle;
   }
 }
