+++ src/adapter/bridges/index.ts
@@ -14,4 +14,3 @@ export * from "./ZKSyncWethBridge";
 export * from "./LineaWethBridge";
 export * from "./BlastBridge";
 export * from "./ScrollERC20Bridge";
-export * from "./ScrollWethBridge";
